Progress on live section
Thought this would be a good stopping point to merge in with master.
Major updates are:
- React component architecture in place for the live page
- Working history and history events sections within the live page
(animates over time to reveal an event outcome). To trigger an event,
you can uncomment lines 27-68 in live-nba-store.js then visit
/frontend/live/. You'll see three events occur, two for your lineup and
one for the opponent.
- Fully tested. 100% coverage except for an if statement related to
Lockr (localStorage), which I'm still trying to think of a way to test
without a browser...thoughts? I see that Lockr doesn't have JS tests to
look at, unfortunately.
- Includes helper methods to convert 8 bit bytes to integers, and you
can pass in whether to make the integer from 2 or 4 bytes (as Coderden
intends to pass us both). We can move this out to its own mixin when we
need to.
- Updated npm modules to their latest versions using npm-check. You'll
notice when running node commands that there may be a few deprecation
warnings. According to the related github issue, should be gone in a
couple weeks https://github.com/postcss/autoprefixer/issues/500